// src/shared/api/generated/model/invitationResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationResponseDtoRole } from './invitationResponseDtoRole.ts';

export interface InvitationResponseDto {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  role: InvitationResponseDtoRole;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$ */
  expiresAt: string;
  code: string;
}